Add tests of Doubly Linked List implementation
For now many of those tests have been marked as failing. Since I wanted
to try out TDD, I decided this was a good moment to do so. I had a first
iteration of the Doubly Linked List implementation, but I want to change
quite a few things on how it works, to make it similar to the current
implementation of the Singly Linked List.

To try TDD, I decided to write the tests of the functionality this
implementation should have, before I go in and refactor the code.