Support negative indexes operations on DLL
Since Doubly Linked Lists keep a reference to the previous element, we
can implement, and make use of negative indexes to insert, or remove, or
traverse the list. This commit makes it so all the methods that take an
index to perform an operation can take a negative index to start looping
from the end (in which index -1 would be equal to list.length-1), and
adds the test cases that illustrate this functionality.