Improve tests to cover traverseToIndex optimization
The traverseToIndex method, on the Doubly Linked List data structure,
thanks to the elements keeping a reference to the previous element, can
have a small optimization to fetch the index faster.

Basically what we do is check the length of elements. If the index we
are trying to access is more than half the length of the linked list,
then we start the looping from the end. For instance, if we had 4
elements, and we were trying to get index 3, or -2 (same element), the
code would realise index 3 falls in the second half of the linked list
length, so it would start looping from the end since it would get there
faster.

Tests have been updated to cover this case. We don't have a direct way
of checking from the test that this optimization was used, but we can
pretty much confirm it was used since nyc, the code coverage tool I use,
will detect the corresponding line that makes that optimization was run.